Alias Kedro IPython extension to kedro.ipython (#1837)
